pyproject.toml
[tool.poetry]
name = "zltk"
version = "0.1.0"
description = ""
authors = ["Zack <zack@zacharymaril.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pypdf2 = "^3.0.1"
openai = "^1.43.0"
spacy = "^3.7.6"
langdetect = "^1.0.9"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
sqlmodel = "^0.0.22"
nltk = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
